Pass build on 32-bit Linux, AsSigned and AsUnsigned are merged into ConvertSign

+++ src/libstd/sys/unix/backtrace.rs
@@ -177,7 +177,7 @@ pub fn write(w: &mut Write) -> io::Result<()> {
         if !ip.is_null() && ip_before_insn == 0 {
             // this is a non-signaling frame, so `ip` refers to the address
             // after the calling instruction. account for that.
-            ip = (ip - 1) as *mut _;
+            ip = (ip as usize - 1) as *mut _;
         }
 
         // dladdr() on osx gets whiny when we use FindEnclosingFunction, and
